+++ geotrek/common/mixins.py
@@ -12,9 +12,9 @@ from django.template.defaultfilters import slugify
 from easy_thumbnails.alias import aliases
 from easy_thumbnails.exceptions import InvalidImageFormatError
 from easy_thumbnails.files import get_thumbnailer
+from embed_video.backends import detect_backend
 
 from geotrek.common.utils import classproperty
-
 
 logger = logging.getLogger(__name__)
 
@@ -156,6 +156,43 @@ class PicturesMixin(object):
             return None
         return os.path.join(settings.MEDIA_URL, th.name)
 
+    @property
+    def videos(self):
+        all_attachments = self.attachments.order_by('-starred')
+        return all_attachments.exclude(attachment_video='')
+
+    @property
+    def serializable_videos(self):
+        serialized = []
+        for att in self.videos:
+            video = detect_backend(att.attachment_video)
+            serialized.append({
+                'author': att.author,
+                'title': att.title,
+                'legend': att.legend,
+                'backend': type(video).__name__.replace('Backend', ''),
+                'code': video.code,
+            })
+        return serialized
+
+    @property
+    def files(self):
+        all_attachments = self.attachments.order_by('-starred')
+        all_attachments = all_attachments.exclude(attachment_file='')
+        return [a for a in all_attachments if not a.is_image]
+
+    @property
+    def serializable_files(self):
+        serialized = []
+        for att in self.files:
+            serialized.append({
+                'author': att.author,
+                'title': att.title,
+                'legend': att.legend,
+                'url': att.attachment_file.url,
+            })
+        return serialized
+
 
 class BasePublishableMixin(models.Model):
     """ Basic fields to control publication of objects.
@@ -167,6 +204,8 @@ class BasePublishableMixin(models.Model):
     publication_date = models.DateField(verbose_name=_(u"Publication date"),
                                         null=True, blank=True, editable=False,
                                         db_column='date_publication')
+    review = models.BooleanField(verbose_name=_(u"Waiting for publication"),
+                                 default=False, db_column='relecture')
 
     class Meta:
         abstract = True
@@ -247,6 +286,8 @@ class PublishableMixin(BasePublishableMixin):
                                                               self.name)
         if self.published:
             s = u'<span class="badge badge-success" title="%s">&#x2606;</span> ' % _("Published") + s
+        elif self.review:
+            s = u'<span class="badge badge-warning" title="%s">&#x2606;</span> ' % _("Waiting for publication") + s
         return s
 
     @property
@@ -329,6 +370,12 @@ class PictogramMixin(models.Model):
 
     def get_pictogram_url(self):
         return self.pictogram.url if self.pictogram else None
+
+
+class OptionalPictogramMixin(PictogramMixin):
+    class Meta:
+        abstract = True
+OptionalPictogramMixin._meta.get_field('pictogram').blank = True
 
 
 class AddPropertyMixin(object):