+++ geotrek/core/tests/test_views.py
@@ -39,7 +39,7 @@ class PathViewsTest(CommonTest):
             'comments': '',
             'departure': '',
             'arrival': '',
-            'datasource': '',
+            'source': '',
             'valid': 'on',
             'geom': '{"geom": "LINESTRING (99.0 89.0, 100.0 88.0)", "snap": [null, null]}',
         }